Fixup premis handling
We have to react to a weirdness in archivematica/issues#743.
We have a couple of options open to us. One way is to test for a
tuple, and if premisrw returns a tuple at any time, we can ignore
it. The next value returned will be the information that we are
interested in. There are notes in the related issue.