Refactor package parsing code a little
There's still a little work to be done here. there's a lot of data
we're passing about between functions both in the call, and the
return values.

We've added some tests to start helping here as well.

A note on fixtures: Apparently I hadn't committed any of the
previously created test fixtures. So I've done that here.