Introduce a new storage service package type
With the amount of data we were passing about and the shape of that
data plus the functions we wanted to perform on it - it started to
look like a type of its own. We convert the data to a type here and
use that in the code itself.

Tests have been added to support the new functionality.

Information is now extracted about all other package types (SIP, DIP)
as well as deleted and replicated packages. There is more info to
be accessed though if we decide we want it, e.g. related packages.