Add AIPs by file format and version reports
This commit adds reports that take a file format name or PUID as input
(to specify a file format or file format version, respectively) and
return information about which AIPs contain files in that format or
format version, as well as their count and size per AIP.

Additionally, this commit reworks the _get_storage_service logic in the
Data and Reporter blueprints to ensure that there are not discrepencies
between the Storage Service name and data returned by Data endpoints.
In Reporter, this commit adds fallback logic similar to what was
previously found in Data. If an invalid Storage Service ID is passed to
the reports view via request parameters, the new helper function will
return one of (in order of preference):
- The requested Storage Service
- The first default Storage Service
- The first Storage Service
- None

This ensures that a valid Storage Service instance is passed along to
the template if one is available. Since it's clearly indicated in the
template UI which Storage Service is currently selected, this component
does not share the issues identified with the previous approach within
the Data blueprint.

Finally, this commit standardizes request parameters in the Reporter
blueprint by adding a shared request_params dictionary and revising
some of the parameters used in the report template's JavaScript to
remove discrepencies.