WIP: Add pie chart for AIPs by format
Experimentation of server-side rendering of pie charts with Plotly's
Python bindings.

Initial thoughts are that the library's a bit difficult to work with
and introduces some heavy-ish dependencies (Plotly, pandas, numpy). But
this might also offer some advantages over making the charts on the
frontend, such as internationalization of labels. Tbd.