Enable integration tests via App Factory
This commit enables use of an in-memory test database by:

- Implementing the Application Factory pattern.
- Improving app config and introducing dev and test configs.
- Adding an `app_instance` fixture which can be used to provide
application context to tests that require use of an in-memory test
database.

It also modifies the installation instructions for the AIPscan and
Celery services, and removes the wsgi.py entrypoint in favour of having
gunicorn call the app factory function directly.