Create a report for user-agents and transfers
We include a data endpoint for retrieving information about users and
the transfers that they are responsible for. Using that data we
construct a tabular log of transfers as well as a prototype timeline
that shows the same information in an interactive Plotly Express
chart. These new reports follow some of the new best practices for
testing and templates established by Tessa in previous commits.

Plotly express is introduced here as well as a complete Fontawesome
package for use in future reports.