Prevent duplicate AIP imports
Previously, each new Fetch Job would result in duplicate AIP, file, and
event records in AIPscan. By calculating and storing a SHA-256 hash for
METS files, this commit removes this duplication. Existing AIP records
are left as-is on subsequent Fetch Jobs if the hash of a downloaded
METS match a stored value. If an AIP is changed, the old records are
deleted from the database prior to creating new AIP and file objects.

Connected to #6