Add Storage Location to reports UI
This commit adds a storage location dropdown selector to the AIPs and
Reports pages.

This commit also updates the reports template to populate file formats
and PUIDs from a Storage Location if one is selected. If no location is
selected, they will continue to be based on the Storage Service.

Each individual report template has also been adjusted to display
Storage Location in addition to Storage Service in the report header,
and dropdown selectors within reports are adjusted to account for
storage locations.