Rework AIP Contents report
This commit modifies the AIP Contents report to a new condensed and
print-friendly format that collapses all of the PUIDs/format versions
for each AIP into a single "Formats" column. For the web UI each entry
is separated by a newline. For the CSV export each entry is separated
by a pipe.

In addition, start and end date filters are added and test coverage for
the Data endpoint improved.