Fix access to PREMIS object
`File.premis_object` has become a `Text.with_variant` to support database
portability (using `LONGTEXT` on MySQL and standard `TEXT` on SQLite). To handle
the new type and possible inconsistencies, this commit adds a new method
`get_premis_object_text()` to the `File` model to normalize the value to a
consistent UTF-8 string.